Rename misc.ts to utils.ts (#2584)
