+++ packages/client/src/game/reducers/initialStates/initialState.ts
@@ -1,5 +1,5 @@
 import { getVariant } from "@hanabi/data";
-import { initArray } from "../../../misc";
+import { initArray } from "../../../utils";
 import * as deckRules from "../../rules/deck";
 import GameMetadata from "../../types/GameMetadata";
 import State from "../../types/State";
