+++ packages/client/src/game/ui/backToLobby.ts
@@ -1,5 +1,5 @@
 import * as chat from "../../chat";
-import { setBrowserAddressBarPath } from "../../misc";
+import { setBrowserAddressBarPath } from "../../utils";
 import globals from "./globals";
 import * as konvaTooltips from "./konvaTooltips";
 import * as timer from "./timer";
