+++ packages/client/src/game/ui/keyboard.ts
@@ -4,8 +4,8 @@ import { parseIntSafe } from "@hanabi/data";
 import * as KeyCode from "keycode-js";
 import Konva from "konva";
 import Screen from "../../lobby/types/Screen";
-import { copyStringToClipboard } from "../../misc";
 import { closeModals, isModalVisible, showPrompt } from "../../modals";
+import { copyStringToClipboard } from "../../utils";
 import * as clueTokensRules from "../rules/clueTokens";
 import * as deckRules from "../rules/deck";
 import ActionType from "../types/ActionType";
