+++ packages/client/src/lobby/login.ts
@@ -4,7 +4,7 @@ import { VERSION } from "@hanabi/data";
 import * as KeyCode from "keycode-js";
 import { FADE_TIME } from "../constants";
 import globals from "../globals";
-import { getURLFromPath, isDevWebpack, isEmpty } from "../misc";
+import { getURLFromPath, isDevWebpack, isEmpty } from "../utils";
 import websocketInit from "../websocketInit";
 import * as nav from "./nav";
 import tablesDraw from "./tablesDraw";
