+++ packages/client/src/lobby/playerSettings.ts
@@ -2,8 +2,8 @@
 
 import { parseIntSafe } from "@hanabi/data";
 import globals from "../globals";
-import { isKeyOf } from "../misc";
 import * as notifications from "../notifications";
+import { isKeyOf } from "../utils";
 
 export function init(): void {
   $("#settings-volume-slider").change(function settingsVolumeSliderChange(
