+++ packages/client/src/lobby/pregame.ts
@@ -3,9 +3,9 @@
 import { getVariant } from "@hanabi/data";
 import * as chat from "../chat";
 import globals from "../globals";
-import { setBrowserAddressBarPath, timerFormatter } from "../misc";
 import * as tooltips from "../tooltips";
 import { OptionIcons } from "../types/OptionIcons";
+import { setBrowserAddressBarPath, timerFormatter } from "../utils";
 import * as nav from "./nav";
 import tablesDraw from "./tablesDraw";
 import Screen from "./types/Screen";
