+++ packages/client/src/lobby/url.ts
@@ -1,6 +1,6 @@
 import { DEFAULT_VARIANT_NAME, parseIntSafe } from "@hanabi/data";
 import globals from "../globals";
-import { setBrowserAddressBarPath } from "../misc";
+import { setBrowserAddressBarPath } from "../utils";
 import WelcomeData from "./types/WelcomeData";
 
 export function parseAndGoto(data: WelcomeData): void {
